bt510adv/requirements.txt
aioserial
greengrasssdk==1.5.0
